requirements.txt
torch
imageio[ffmpeg]
pyyaml
tqdm
matplotlib
numpy==1.26.4